fix(cont): A8 续接 OCE 单独分支 + 2026-09-05 评审 75 项修复核对落盘（290/290 全绿）
- A8：OutputContinuer 续接请求 catch(Exception) 拆分 OperationCanceledException 单独分支——流式向上抛（PumpResponseAsync 记响应管道异常并关闭连接）、非流式 break 收尾，续接取消/超时不再误判 bad_alloc 崩溃恢复
- 2026-09-05 评审报告 75 项逐行回源核对（P0 6/6、P1 8/9），核对报告 + 汇总图落盘 docs/
- 适配 9 个修复未同步测试：TokenGuard backend 非空校验（NeverBackend stub）、ExtraArgs 引号剔除断言、日志轮切时间戳命名断言
- 架构设计说明书 v2.26.2 → v2.26.3

+++ LlamaHarness/PerfEventTracker.cs
@@ -1,7 +1,11 @@
 namespace LlamaHarness;
 
-/// <summary>性能事件会话聚合（按 Category+Op 归组）。</summary>
-public sealed record PerfEventStats(int Count, double SumMs, double MaxMs)
+/// <summary>
+/// 性能事件会话聚合（按 Category+Op 归组）。
+/// 问题 19 修复：Count 由 int → long。长期运行（7×24 高频事件如 slot_select/wakeup）
+/// int（约 21 亿）存在溢出风险，溢出后 AvgMs 会因负数返回异常值。
+/// </summary>
+public sealed record PerfEventStats(long Count, double SumMs, double MaxMs)
 {
     /// <summary>平均耗时（ms）。</summary>
     public double AvgMs => Count > 0 ? SumMs / Count : 0;
